requirements.txt
ocs_academic_hub==0.99.33
pandas==1.3.1
python_dateutil==2.8.2
pytimeparse==1.1.8
requests==2.26.0
streamlit==1.3.0
urllib3==1.26.6
